requirements.txt
dash>=2.14,<4
plotly>=5.18,<6
dash-bootstrap-components>=1.5,<2
dash-ag-grid>=31.0,<32
snowflake-connector-python>=3.6,<5
pandas>=2.0,<3
numpy>=1.24,<3
pyarrow>=14,<24
python-pptx>=1.0,<2
flask-caching>=2.0,<3
python-dotenv>=1.0,<2
kaleido>=0.2,<1
